#7553 Refactor
