+++ packages/ketcher-react/src/script/ui/action/fullscreen.ts
@@ -15,6 +15,7 @@
  ***************************************************************************/
 
 import isHidden from './isHidden';
+import { getFullscreenElement } from '../../../utils';
 
 const requestFullscreen = (element: HTMLElement) => {
   (element.requestFullscreen && element.requestFullscreen()) ||
@@ -37,48 +38,6 @@ const getIfFullScreen = () => {
     document.webkitFullscreenElement ||
     document.msFullscreenElement
   );
-};
-
-const getFullscreenElement = (): HTMLElement => {
-  const ketcherRoot = document.querySelector(
-    '.Ketcher-root, .Ketcher-polymer-editor-root',
-  ) as HTMLElement;
-
-  if (ketcherRoot) {
-    const markedContainer = ketcherRoot.closest(
-      '[data-ketcher-fullscreen-container]',
-    );
-    if (markedContainer) {
-      return markedContainer as HTMLElement;
-    }
-
-    const editorRoot = ketcherRoot.closest('[data-ketcher-editor]');
-    if (editorRoot) {
-      const editorParent = editorRoot.parentElement;
-      if (
-        editorParent &&
-        editorParent !== document.body &&
-        editorParent !== document.documentElement
-      ) {
-        if (editorParent.id === 'root') {
-          return editorParent;
-        }
-        return editorParent;
-      }
-
-      const rootElement = editorRoot.closest('#root');
-      if (rootElement) {
-        return rootElement as HTMLElement;
-      }
-    }
-
-    const rootById = document.getElementById('root');
-    if (rootById) {
-      return rootById;
-    }
-  }
-
-  return document.getElementById('root') || document.documentElement;
 };
 
 const toggleFullscreen = () => {
